services/ai_orchestrator/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
pydantic==2.5.0
redis==5.0.1
tenacity
google-auth
